calendar 작업 거의 완료

+++ src/main/java/com/example/world/calendar/CalendarRepository.java
@@ -2,6 +2,6 @@ package com.example.world.calendar;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CalendarRepository extends JpaRepository<Calendar, Long> {
+public interface CalendarRepository extends JpaRepository<Calendar, Integer> {
 
 }
